Use size_t instead of int in SpellParser::PositionPair
Use of int required a lot of casts when using the indices
with std::string methods. Furthermore code used -1 as
error code, which should have used std::optional for invalid
PoistionPair instnaces.

Test-Information:

All tests pass on Debian Stretch with clang-3.9.

+++ SwifTools/MacOSXChecker.mm
@@ -69,7 +69,7 @@ void MacOSXChecker::checkFragment(const std::string& fragment, PositionPairList&
         if (range.location != NSNotFound) {
             if (range.location < nextLocation)
                 break;
-            misspelledPositions.push_back(PositionPair(static_cast<int>(range.location), static_cast<int>(range.location + range.length)));
+            misspelledPositions.push_back(PositionPair(range.location, range.location + range.length));
             nextLocation = range.location + range.length + 1;
         }
     } while (range.location != NSNotFound);
